feat: tile types exist now, and every tile has one upon creation
