+++ src/main/java/com/cozystudios/fofrathzindustry/TileType.java
@@ -1,0 +1,10 @@
+package com.cozystudios.fofrathzindustry;
+
+public enum TileType
+{
+    Grass,
+    //Stone,
+    //Copper,
+    //Iron,
+    Coal
+}
